TechWorldWebParent/TechWorldBackEnd/src/main/java/com/techworld/admin/category/CategoryRepository.java
package com.techworld.admin.category;

import com.techworld.common.entity.Category;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Modifying;
import org.springframework.data.jpa.repository.Query;
import org.springframework.stereotype.Repository;

import java.util.List;

@Repository
public interface CategoryRepository extends JpaRepository<Category, Integer> {
    public Long countById(Integer id);

    public Category findByName(String name);

    @Query("update Category  c set c.enabled = ?2 where c.id = ?1")
    @Modifying
    public void updateEnabledStatus(Integer id, boolean enabled);
}

